+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ## [Unreleased]
 - #task_number - short task description
 - #2 - entities personal_address, pickup_point, postomat where created, integrated into gazon-service and tested
-- #4 - entity BankCard was created, integrated into gazon-service and tested
 - #3 - created passport entity
+- #4 - entity BankCard was created, integrated into gazon-service and tested
+- #6 - added pagination for Example
+- #7 - entities Product and ProductImage were created and integrated
 
 ## [1.0.0]